Reduce named pipe server boilerplate
